move data fetching to parent page.tsx component
